+++ src/OndaEDF.jl
@@ -21,6 +21,4 @@ export extract_channels_by_label, edf_signals_to_samplesinfo
 include("export_edf.jl")
 export onda_to_edf
 
-include("prettyprint_diagnostic_info.jl")
-
 end # module
