+++ src/import_edf.jl
@@ -153,6 +153,7 @@ function edf_signal_encoding(edf_signal_header, edf_seconds_per_record)
             sample_type=sample_type)
 end
 
+# TODO: replace this with float type for mismatched
 function promote_encodings(encodings; pick_offset=(_ -> 0.0), pick_resolution=minimum)
     sample_type = reduce(promote_type, (e.sample_type for e in encodings))
 
@@ -189,47 +190,6 @@ end
 ##### `EDF.Signal`s -> `Onda.Samples`
 #####
 
-function extract_channels(edf_signals, channel_matchers)
-    extracted_channels = Pair{String,EDF.Signal}[]
-    for channel_matcher in channel_matchers
-        for edf_signal in edf_signals
-            edf_signal isa EDF.Signal || continue
-            # skip channels already extracted
-            any(x -> last(x) === edf_signal, extracted_channels) && continue
-            channel_name = channel_matcher(edf_signal)
-            channel_name === nothing && continue
-            push!(extracted_channels, channel_name => edf_signal)
-        end
-    end
-    return extracted_channels
-end
-
-"""
-    edf_signals_to_samplesinfo(edf, edf_signals, kind, channel_names, samples_per_record; unit_alternatives=STANDARD_UNITS)
-
-Generate a single `Onda.SamplesInfo` for the given collection of `EDF.Signal`s
-corresponding to the channels of a single Onda signal.  Sample units are
-converted to Onda units and checked for consistency, and a promoted encoding
-(resolution, offset, and sample type/rate) is generated.
-
-No conversion of the actual signals is performed at this step.
-"""
-function edf_signals_to_samplesinfo(edf::EDF.File, edf_signals::Vector{<:EDF.Signal}, kind, channel_names; unit_alternatives=STANDARD_UNITS)
-    onda_units = map(s -> edf_to_onda_unit(s.header.physical_dimension, unit_alternatives), edf_signals)
-    onda_sample_unit = first(onda_units)
-
-    edf_encodings = map(s -> edf_signal_encoding(s.header, edf.header.seconds_per_record), edf_signals)
-    onda_encoding = promote_encodings(edf_encodings)
-
-    info = SamplesInfo(; kind=kind, channels=channel_names,
-                       sample_unit=string(onda_sample_unit),
-                       sample_resolution_in_unit=onda_encoding.sample_resolution_in_unit,
-                       sample_offset_in_unit=onda_encoding.sample_offset_in_unit,
-                       sample_type=onda_encoding.sample_type,
-                       sample_rate=onda_encoding.sample_rate)
-    return info
-end
-
 struct SamplesInfoError <: Exception
     msg::String
     cause::Exception
@@ -246,83 +206,6 @@ function groupby(f, list)
         push!(get!(d, f(v), Vector{eltype(list)}()), v)
     end
     return d
-end
-
-"""
-    extract_channels_by_label(edf::EDF.File, signal_names, channel_names;
-                              unit_alternatives=STANDARD_UNITS, preprocess_labels=(l,t) -> l)
-
-For one or more signal names and one or more channel names,
-return a list of `[(infos, errors)...]` where `errors` is a vector of errors that occurred,
-and `infos` is a vector of `(si::Onda.SamplesInfo, edf_signals::Vector{EDF.Signal})`,
-where `edf_signals` align with `si.channel`. This list can have more than one pair
-if channels with the same signal kind/type have different sample rates or
-physical units.
-
-`(label, transducer_type)` pairs will be transformed into labels by `preprocess_labels`
-(default preprocessor returns the original label). An alternate to `STANDARD_UNITS`
-for mapping different spellings of units to their canonical unit name can be passed in
-as `unit_alternatives`.
-
-`errors` contains `SamplesInfoError`s thrown if channels corresponding to a signal
-were extracted but an error occured while interpreting physical units,
-while promoting sample encodings, or otherwise constructing a `SamplesInfo`.
-
-`signal_names` should be an iterable of `String`s naming the signal types to
-extract (e.g., `["ecg", "ekg"]`; `["eeg"]`).
-
-`channel_names` should be an iterable of channel specifications, each of which
-can be either a `String` giving the generated channel name, or a `Pair` mapping
-a canonical name to a list of alternatives that it should be substituted for
-(e.g., `"canonical_name" => ["alt1", "alt2", ...]`).
-
-`unit_alternatives` lists standardized unit names and alternatives that map to them.
-See `OndaEDF.STANDARD_UNITS` for defaults.
-
-`preprocess_labels(label::String, transducer_type::String)` is applied to raw edf signal header labels
-beforehand; defaults to returning `label`.
-
-See `OndaEDF.STANDARD_LABELS` for the labels (`signal_names => channel_names`
-`Pair`s) that are used to extract EDF signals by default.
-
-"""
-function extract_channels_by_label(edf::EDF.File, signal_names, channel_names;
-                                   unit_alternatives=STANDARD_UNITS, preprocess_labels=(l,t) -> l)
-    matcher = x -> begin
-        # yo I heard you like closures
-        # x is either a channel name (string), or a channel_name => alternatives Pair.
-        this_channel_name = x isa Pair ? first(x) : x
-        return s -> match_edf_label(preprocess_labels(s.header.label, s.header.transducer_type),
-                                    signal_names,
-                                    this_channel_name,
-                                    channel_names)
-    end
-    edf_channels = extract_channels(edf.signals, (matcher(x) for x in channel_names))
-
-    # place channels with different physical units or sample rates in separate Onda signals
-    grouped = groupby(edf_channels) do p
-        channel_name, edf_signal = p
-        return ((edf_signal.header.samples_per_record / edf.header.seconds_per_record), edf_signal.header.physical_dimension)
-    end
-
-    results = map(values(grouped)) do pairs
-        # pairs is a vector of `standard_onda_name::String => EDF.Signal` pairs
-        edf_channel_names, edf_channels = zip(pairs...)
-        try
-            edf_channels = collect(edf_channels)
-            info = edf_signals_to_samplesinfo(edf, edf_channels, first(signal_names), collect(edf_channel_names))
-            return info, edf_channels
-        catch e
-            # do not throw, but return any errors
-            units = [s.header.label => s.header.physical_dimension for s in edf_channels]
-            msg ="""Skipping signal: error while processing units and encodings
-                    for $(first(signal_names)) signal with units $units"""
-            return SamplesInfoError(msg, e)
-        end
-    end
-
-    return ([info_channels for info_channels in results if !isa(info_channels, Exception)],
-            [e for e in results if isa(e, Exception)])
 end
 
 # "channel"
@@ -548,9 +431,6 @@ following transformations:
   (e.g. for an EOG matched channel, "eogl", "loc", "lefteye", etc. are converted
   to "left").
 
-`EDF.Signal`s which get extracted into more than one `Onda.Samples` are removed
-and an `AmbiguousChannelError` displayed as a warning.
-
 See the OndaEDF README for additional details regarding EDF formatting expectations.
 """
 function store_edf_as_onda(edf::EDF.File, onda_dir, recording_uuid::UUID=uuid4();
@@ -626,33 +506,6 @@ function validate_arrow_prefix(prefix)
     return prefix
 end
 
-# the same EDF.Signal was extracted into multiple `Onda.SampleInfos`.
-struct AmbiguousChannelError <: Exception
-    summary
-end
-
-Base.showerror(io::IO, e::AmbiguousChannelError) = print(io, "AmbiguousChannelError: the same `EDF.Signal` was extracted into multiple `Onda.SampleInfo`s\n", e.summary)
-
-# return a list of tuples describing sample info channels that are ambiguous
-# because they were extracted from the same edf_signal
-# modifies `infos` and associated `edf_signals::Vector{EDF.Signal}` in place to remove ambiguous signals
-function _ambiguous_channels!(extracted, edf_signal)
-    ambiguous = []
-    for (info, edf_signals) in extracted
-        i = findfirst(==(edf_signal), edf_signals)
-        isnothing(i) && continue
-        channel_name = info.channels[i]
-        summary = (info.kind, channel_name, info.sample_unit)
-        push!(ambiguous, (info, edf_signals) => summary)
-    end
-    length(ambiguous) < 2 && return nothing
-    for ((info, edf_signals), (_, channel_name, _)) in ambiguous
-        filter!(!=(channel_name), info.channels)
-        filter!(!=(edf_signal), edf_signals)
-    end
-    return map(last, ambiguous)
-end
-
 """
     edf_to_onda_samples(edf::EDF.File; custom_extractors=())
 
@@ -687,125 +540,6 @@ function edf_to_onda_samples(edf::EDF.File; kwargs...)
     samples, exec_plan = execute_plan(signals_plan, edf)
     return samples, exec_plan
 end
-
-"""
-    edf_header_to_onda_samples_info(edf::EDF.File; custom_extractors=STANDARD_EXTRACTORS)
-
-Read edf header, return a mapping from `Onda.SamplesInfo`s to the vector
-of `EDF.Signals` it was extracted from, along with a `NamedTuple` containing
-diagnostic information in fields:
-- `header_map`: a vector of corresponding `Onda.SamplesInfo`, `Vector{EDF.SignalHeader}` pairs.
-- `unextracted_edf_headers`: a vector of EDF signal headers that could not be extracted.
-- `errors`: vector of exceptions thrown while attempting to extract header info.
-
-The `NamedTuple` can be pretty-printed and used to compare outputs with `SamplesInfo`s
-previously extracted from the same data, for testing purposes.
-
-`EDF.read!` does not get called, this function will work
-with only the first few bytes--30k should be enough--of the edf file,
-which is convenient for developping custom extractors for a dataset without
-reading and converting all the samples data.
-"""
-function edf_header_to_onda_samples_info(edf::EDF.File; custom_extractors=STANDARD_EXTRACTORS)
-    info_map, errors = try
-        matched = []
-        errors = Exception[]
-        for extractor in custom_extractors
-            extracteds, errs = extractor(edf)
-            append!(errors, errs)
-            for extracted in extracteds
-                push!(matched, extracted)
-            end
-        end
-        # each edf_signal should get extracted into at most one SamplesInfo, otherwise it is ambiguous
-        matched_signals = Set(Iterators.flatten(map(last, matched)))
-        ambiguous_edf_signals = []
-        for edf_signal in matched_signals
-            ambiguous_channels = _ambiguous_channels!(matched, edf_signal)
-            if !isnothing(ambiguous_channels)
-                edf_signal_summary = (edf_signal.header.label,
-                                      edf_signal.header.transducer_type,
-                                      edf_signal.header.physical_dimension)
-                push!(errors, AmbiguousChannelError(edf_signal_summary => ambiguous_channels))
-            end
-        end
-        matched = [(info, edf_signals) for (info, edf_signals) in matched if !isempty(info.channels)]
-        sort!(errors; by=string)
-        matched, errors
-    catch e
-        [], [e]
-    end
-    matched_edf_headers = mapreduce(last, union, info_map; init=[])
-    unextracted = [s.header for s in edf.signals if isa(s, EDF.Signal) && s ∉ matched_edf_headers]
-    header_map = [info => [s.header for s in edf_signals if isa(s, EDF.Signal)] for (info, edf_signals) in info_map]
-    return info_map, (header_map=header_map,
-                      unextracted_edf_headers=unextracted,
-                      errors=errors)
-end
-
-function diagnostics_from_plan(plan)
-    header_map = []
-    unextracted = []
-    errors = Exception[]
-
-    rows = Tables.rows(plan)
-
-    for row in rows
-        err = _get(row, :error)
-        if err isa Exception
-            push!(errors, err)
-        end
-    end
-
-    foreach(groupby(grouper((:onda_signal_idx, )), rows)) do (_, rows)
-        headers = _signal_header.(rows)
-        try
-            if any(ismissing, _get.(rows, :channel))
-                append!(unextracted, headers)
-            else
-                info = merge_samples_info(rows)
-                push!(header_map, _samples_info(info) => headers)
-            end
-        catch e
-            push!(errors, e)
-            append!(unextracted, headers)
-        end
-    end
-
-    return (; header_map, unextracted, errors)
-end
-
-function _signal_header(row)
-    fields = fieldnames(EDF.SignalHeader)
-    values = NamedTuple{fields}(NamedTuple(row))
-    return EDF.SignalHeader(values...)
-end
-
-function _samples_info(row)
-    fields = (:kind, :channels, :sample_unit, :sample_resolution_in_unit,
-              :sample_offset_in_unit, :sample_type, :sample_rate)
-    values = NamedTuple{fields}(NamedTuple(row))
-    return Onda.SamplesInfo(; values...)
-end
-
-function diagnostics_table(diagnostics)
-    header_map, unextracted_edf_headers, errors = diagnostics
-    diag_table = []
-    for (samplesinfo, headers) in header_map
-        for (header, channel) in zip(headers, samplesinfo.channels)
-            push!(diag_table, (; NamedTuple(samplesinfo)..., channel, _named_tuple(header)...))
-        end
-    end
-
-    for header in unextracted_edf_headers
-        push!(diag_table, _named_tuple(header))
-    end
-
-    return diag_table
-end
-
-
-
 
 """
     edf_to_onda_annotations(edf::EDF.File, uuid::UUID)
